orquestador/pyproject.toml
[tool.poetry]
name = "orquestador"
version = "0.1.0"
description = "Orquestador de jobs de entrenamiento"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "*"
pandas = "*"
scikit-learn = "*"
joblib = "*"
pyyaml = "*"
tqdm = "*"
boto3 = "*"
kubernetes = "*"
pydantic = "*"
typer = "*"
loguru = "*"
reactivex = "*"
kafka-python = "^2.3.0"
